requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
blinker==1.9.0
boto3==1.37.31
botocore==1.37.31
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
databricks-sdk==0.44.1
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docker==7.1.0
dynaconf==3.2.10
evidently==0.7.0
executing==2.2.0
Faker==37.1.0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
fonttools==4.56.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.32.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kafka-python==2.1.5
kiwisolver==1.4.8
litestar==2.15.2
litestar-htmx==0.4.1
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.15
mlflow==2.20.3
mlflow-skinny==2.20.3
msgspec==0.19.0
multidict==6.4.3
multipart==1.2.1
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numba==0.61.2
numpy==2.0.2
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
polyfactory==2.21.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.1
# pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rich-click==1.8.8
rsa==4.9
s3transfer==0.11.4
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
shap==0.47.1
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.1
statsmodels==0.14.4
tenacity==9.1.2
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
ujson==5.10.0
urllib3==2.3.0
uuid6==2024.7.10
uvicorn==0.34.0
waitress==3.0.2
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.0
zipp==3.21.0
